Added logic to the PostController so that users are the only ones who can edit or delete their posts

+++ app/Http/Controllers/PostsController.php
@@ -122,6 +122,14 @@ class PostsController extends Controller
         // The variable that finds the id for the post from the database
         $post = Post::find($id);
 
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+
+            // If user is not authorized to edit the post redirect them back to viewing posts
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+
+        }
+
         // The view for that post to be edited
         return view('posts.edit')->with('post', $post);
 
@@ -168,6 +176,14 @@ class PostsController extends Controller
         
         // The variable that finds the id for the post from the database
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+
+            // If user is not authorized to delete the post redirect them back to viewing posts
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+
+        }
 
         // Deletes the post
         $post->delete();
